Create Obstacle_management.py
This code uses the Haar Cascade Classifier to detect cars in the video feed. It then calculates the distance to the closest car based on the width and height of the bounding box. If the distance is less than 20 meters, the target speed is set to 10 km/h, otherwise, it's set to 20 km/h. It then calculates the error between the target speed and the current speed, and uses a PID controller to determine the control output. The control output is added to the current speed, and the speed is limited to a maximum of 30 km/h. Finally, the current speed and target speed are displayed on the screen, and the program waits for the 'q' key to be pressed to quit. Note that the PID parameters may need to be adjusted